Copy all metadata from file referenced by merge element, not just info
When a merge element references another file, any title, titleabbrev, or
subtitle elements that are *not* inside an info element, as well as the
contents of the info element if present, are copied from the referenced
file into the info element in the current module. The warning that the
referenced file must contain an info element is removed.

This (I think) more closely meets the requirement that "Authors are not
 required to [place all metadata inside an info wrapper] in order to use
assemblies. Processing systems are to behave as if they had."